ref(btx): move spec tests into contrib modules
Convert BTX from a standalone Go module into shared internal test utilities
and run provider-specific spec tests from their owning contrib modules. This
keeps provider SDK dependencies scoped to the integrations they validate and
removes the pinned unreleased module release path.
