+++ internal/btx/span_fetcher.go
