+++ internal/btx/span_validator.go
