+++ internal/btx/testdata/cassettes/anthropic/attachments.yaml
