+++ internal/btx/testdata/cassettes/anthropic/streaming.yaml
