+++ internal/btx/testdata/cassettes/bedrock/attachments.yaml
