+++ internal/btx/testdata/cassettes/bedrock/converse_stream.yaml
