+++ internal/btx/testdata/cassettes/openai/attachments.yaml
