+++ internal/btx/testdata/cassettes/openai/completions.yaml
