+++ internal/btx/testdata/cassettes/openai/reasoning.yaml
