+++ internal/btx/testdata/cassettes/openai/streaming.yaml
