+++ internal/btx/testdata/cassettes/openai/tools.yaml
