--- scripts/pinned_unreleased_modules.txt
@@ -1,14 +1,0 @@
-# Modules that depend on the Braintrust SDK but are not themselves released as
-# tagged Go modules. Their go.mod files are still pinned by prepare_release.sh
-# so they stay in sync with the release version (otherwise mod-verify fails
-# during release prep because workspace-mode tidy bumps their require lines).
-#
-# Add a new entry here when:
-#   - You create a new top-level Go module under the repo (sibling to btx/)
-#   - The module imports github.com/braintrustdata/braintrust-sdk-go or a
-#     nested contrib module
-#   - The module is NOT meant to be published as its own tagged release
-#
-# Format: one repo-relative path per line. Blank lines and lines starting with
-# '#' are ignored.
-btx
